+++ projects/socket_example/src/stm32f0xx/main.c
@@ -1,0 +1,4 @@
+
+int main(void) {
+  return 0;
+}
